design_introspection: Integrate select_to_tcl_list command with design introspection plugin
