+++ design_introspection-plugin/Makefile
@@ -5,6 +5,7 @@ SOURCES = design_introspection.cc \
 	  get_ports.cc \
 	  get_cells.cc \
 	  get_pins.cc \
-	  get_count.cc
+	  get_count.cc \
+	  selection_to_tcl_list.cc
 
 include ../Makefile_plugin.common
