+++ design_introspection-plugin/selection_to_tcl_list.h
@@ -15,28 +15,22 @@
  *  WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
  *  ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
  *  OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
- *
  */
+#ifndef _SELECTION_TO_TCL_LIST_H_
+#define _SELECTION_TO_TCL_LIST_H_
 
-#include "get_cells.h"
-#include "get_count.h"
-#include "get_nets.h"
-#include "get_pins.h"
-#include "get_ports.h"
-#include "selection_to_tcl_list.h"
+#include "kernel/register.h"
 
 USING_YOSYS_NAMESPACE
 
-PRIVATE_NAMESPACE_BEGIN
+struct SelectionToTclList : public Pass {
+    SelectionToTclList() : Pass("selection_to_tcl_list", "Extract selection to TCL list") {}
 
-struct DesignIntrospection {
-    DesignIntrospection() {}
-    GetNets get_nets_cmd;
-    GetPorts get_ports_cmd;
-    GetCells get_cells_cmd;
-    GetPins get_pins_cmd;
-    GetCount get_count_cmd;
-    SelectionToTclList selection_to_tcl_list_cmd;
-} DesignIntrospection;
+    void help() override;
+    void execute(std::vector<std::string> args, RTLIL::Design *design) override;
 
-PRIVATE_NAMESPACE_END
+  private:
+    void AddObjectNameToTclList(RTLIL::IdString &module, RTLIL::IdString &object, Tcl_Obj *tcl_list);
+};
+
+#endif // SELECTION_TO_TCL_LIST_H_
