+++ design_introspection-plugin/tests/selection_to_tcl_list/selection_to_tcl_list.golden.txt
@@ -1,0 +1,3 @@
+{middle/$add$selection_to_tcl_list.v:32$5} top/middle_inst_4 top/middle_inst_3 top/middle_inst_2 top/middle_inst_1 {top/$add$selection_to_tcl_list.v:14$2} top/ibuf_inst top/ibuf_proxy
+{middle/$1\cnt[1:0]} {middle/$add$selection_to_tcl_list.v:32$5_Y} {middle/$0\cnt[1:0]} middle/clk_int middle/cnt middle/out middle/clk {top/$1\cnt[1:0]} {top/$add$selection_to_tcl_list.v:14$2_Y} {top/$0\cnt[1:0]} top/ibuf_out top/ibuf_proxy_out top/clk_int_2 top/clk_int_1 top/cnt top/out top/in top/clk2 top/clk
+{middle/$1\cnt[1:0]} {middle/$add$selection_to_tcl_list.v:32$5_Y} {middle/$0\cnt[1:0]} middle/clk_int middle/cnt middle/out middle/clk {middle/$add$selection_to_tcl_list.v:32$5} {middle/$proc$selection_to_tcl_list.v:28$6} {middle/$proc$selection_to_tcl_list.v:31$4} {top/$1\cnt[1:0]} {top/$add$selection_to_tcl_list.v:14$2_Y} {top/$0\cnt[1:0]} top/ibuf_out top/ibuf_proxy_out top/clk_int_2 top/clk_int_1 top/cnt top/out top/in top/clk2 top/clk top/middle_inst_4 top/middle_inst_3 top/middle_inst_2 top/middle_inst_1 {top/$add$selection_to_tcl_list.v:14$2} top/ibuf_inst top/ibuf_proxy {top/$proc$selection_to_tcl_list.v:6$3} {top/$proc$selection_to_tcl_list.v:13$1}
