--- selection-plugin/Makefile
@@ -1,3 +1,0 @@
-NAME = selection
-SOURCES = selection.cc
-include ../Makefile_plugin.common
