--- selection-plugin/tests/Makefile
@@ -1,3 +1,0 @@
-TESTS = counter
-include $(shell pwd)/../../Makefile_test.common
-counter_verify = $(call diff_test,counter,txt)
